Refactor catalog request form
