--- catalog/ui/src/app/Catalog/Request/catalog-request.css
@@ -1,12 +1,0 @@
-.pf-c-form.rhpds-catalog-request-form {
-  margin-top: 20pt;
-}
-
-.pf-c-form__group.rhpds-catalog__terms-of-service {
-  background-color: var(--pf-global--BackgroundColor--light-300);
-  padding: 12px 16px;
-}
-
-.rhpds-catalog-request-form .pf-c-check {
-  padding-top: 18pt;
-}
