+++ catalog/ui/src/app/Catalog/catalog-request.css
@@ -1,0 +1,13 @@
+.catalog-request-form.pf-c-form {
+  margin-top: 20pt;
+}
+
+.catalog-request-form.pf-c-form .pf-c-check {
+  padding-top: 18pt;
+}
+
+.catalog-terms-of-service.pf-c-form__group {
+  background-color: var(--pf-global--BackgroundColor--light-300);
+  padding: 12px 16px;
+}
+
